Update SQLDiscardGenerator to extract out more duplicates
