+++ src/sqlancer/yugabyte/ysql/gen/YSQLDiscardGenerator.java
@@ -14,25 +14,10 @@ public final class YSQLDiscardGenerator extends SQLDiscardGenerator {
     }
 
     public static SQLQueryAdapter create(YSQLGlobalState globalState) {
-        StringBuilder sb = new StringBuilder();
-        sb.append("DISCARD ");
         // prevent that DISCARD discards all tables (if they are TEMP tables)
         boolean hasNonTempTables = globalState.getSchema().getDatabaseTables().stream()
                 .anyMatch(t -> t.getTableType() == TableType.STANDARD);
-        String what;
-        if (hasNonTempTables) {
-            what = Randomly.fromOptions("ALL", "PLANS", "SEQUENCES", "TEMPORARY", "TEMP");
-        } else {
-            what = Randomly.fromOptions("PLANS", "SEQUENCES");
-        }
-        sb.append(what);
-        return new SQLQueryAdapter(sb.toString(), ExpectedErrors.from("cannot run inside a transaction block")) {
 
-            @Override
-            public boolean couldAffectSchema() {
-                return canDiscardTemporaryTables(what);
-            }
-        };
+        return SQLDiscardGenerator.create(hasNonTempTables);
     }
-
 }
